requirements.txt
pika==0.10.0
PyYAML==3.11
requests==2.10.0
wheel==0.24.0
